Bridge tokens: µs-precision expiry so same-second tokens are unique
Two back-to-back adapter fetches minted identical payloads (integer expiry +
same user/purpose → same HMAC), so the second legitimately-minted token was
refused as a replay. Both minters emit float expiry; both validators parse float.

+++ dashboard/bridge.py
@@ -63,7 +63,9 @@ def validate_bridge_token(raw: str) -> str | None:
         return None
     _, expiry_s, user, purpose = parts
     try:
-        expiry = int(expiry_s)
+        # float (µs precision): tokens minted in the same second must still be
+        # unique, or back-to-back requests trip the single-use replay guard
+        expiry = float(expiry_s)
     except ValueError:
         return None
     now = time.time()
